ref: Moved filter, content item, and content from catalogue module to separate folder
